Refactor User and compiled jpa
