+++ build.gradle
@@ -37,7 +37,7 @@ dependencies {
 
     compile 'org.webjars:jquery-ui:1.11.4'
 
-// https://mvnrepository.com/artifact/org.springframework.data/spring-data-elasticsearch
+    // https://mvnrepository.com/artifact/org.springframework.data/spring-data-elasticsearch
     compile group: 'org.springframework.data', name: 'spring-data-elasticsearch', version: '3.2.0.RELEASE'
 
     // https://mvnrepository.com/artifact/org.springframework.boot/spring-boot-starter-thymeleaf
@@ -45,10 +45,15 @@ dependencies {
 
     // https://mvnrepository.com/artifact/org.springframework.security/spring-security-oauth2-client
     compile group: 'org.springframework.security', name: 'spring-security-oauth2-client', version: '5.2.0.RELEASE'
+
     // https://mvnrepository.com/artifact/org.springframework.security/spring-security-oauth2-jose
     compile group: 'org.springframework.security', name: 'spring-security-oauth2-jose', version: '5.2.0.RELEASE'
-// https://mvnrepository.com/artifact/org.springframework.security.oauth.boot/spring-security-oauth2-autoconfigure
+
+    // https://mvnrepository.com/artifact/org.springframework.security.oauth.boot/spring-security-oauth2-autoconfigure
     compile group: 'org.springframework.security.oauth.boot', name: 'spring-security-oauth2-autoconfigure', version: '2.1.7.RELEASE'
+
+    // https://mvnrepository.com/artifact/org.springframework.boot/spring-boot-starter-data-jpa
+    compile group: 'org.springframework.boot', name: 'spring-boot-starter-data-jpa', version: '2.2.1.RELEASE'
 
 
 }
