+++ src/main/java/com/lecturesearch/lecture/HomeController.java
@@ -7,6 +7,7 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.data.web.PageableDefault;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
+import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 
 import javax.servlet.http.HttpServletResponse;
@@ -24,6 +25,11 @@ public class HomeController {
 //        model.addAttribute("list", rsList);
 //        return "/layout/main";
 //    }
+
+    @GetMapping("/")
+    public String home(){
+        return "main";
+    }
 
     @RequestMapping("/main")
     public String list(@PageableDefault Pageable pageable, Model model, HttpServletResponse response) {
