+++ src/main/java/com/lecturesearch/lecture/OAuth2/controller/UserController.java
@@ -1,6 +1,6 @@
 package com.lecturesearch.lecture.OAuth2.controller;
 
-import com.lecturesearch.lecture.OAuth2.User;
+import com.lecturesearch.lecture.OAuth2.domain.User;
 import com.lecturesearch.lecture.OAuth2.annotation.SocialUser;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -16,13 +16,13 @@ public class UserController {
 
     @GetMapping(value = "/loginSuccess")
     public String loginComplete(@SocialUser User user) {
-        System.out.println("email : "+user.getEmail()+" name : "+user.getName());
+        System.out.println("email : "+user.getEmail()+"  | name : "+user.getName());
         return "redirect:/main";
     }
 
     @PostMapping("/create")
     public String create(User user){
-        System.out.println("email : "+user.getEmail()+" name : "+user.getName()+" password : "+user.getPassword());
+        System.out.println("email : "+user.getEmail()+"  |  name : "+user.getName()+"  | password : "+user.getPassword());
         return "redirect:/main";
     }
