+++ src/main/java/com/lecturesearch/lecture/OAuth2/repository/UserRepository.java
@@ -1,6 +1,6 @@
 package com.lecturesearch.lecture.OAuth2.repository;
 
-import com.lecturesearch.lecture.OAuth2.User;
+import com.lecturesearch.lecture.OAuth2.domain.User;
 import org.springframework.data.elasticsearch.repository.ElasticsearchRepository;
 
 public interface UserRepository extends ElasticsearchRepository<User, String> {
