+++ src/main/java/com/lecturesearch/lecture/OAuth2/resolver/UserArgumentResolver.java
@@ -2,7 +2,7 @@ package com.lecturesearch.lecture.OAuth2.resolver;
 
 
 import com.lecturesearch.lecture.OAuth2.SocialType;
-import com.lecturesearch.lecture.OAuth2.User;
+import com.lecturesearch.lecture.OAuth2.domain.User;
 import com.lecturesearch.lecture.OAuth2.repository.UserRepository;
 import com.lecturesearch.lecture.OAuth2.annotation.SocialUser;
 import org.springframework.core.MethodParameter;
@@ -21,7 +21,6 @@ import org.springframework.web.method.support.ModelAndViewContainer;
 
 import javax.servlet.http.HttpSession;
 import java.time.LocalDateTime;
-import java.util.HashMap;
 import java.util.Map;
 
 import static com.lecturesearch.lecture.OAuth2.SocialType.FACEBOOK;
