+++ src/main/resources/templates/form.html
@@ -39,12 +39,12 @@
             <div class="form-group">
                 <table style="text-align: center; align-content: center; margin: auto; margin-bottom: 20px;">
                     <tr>
-                        <td style="width: 110px;"><label for="loginEmail">이메일</label></td>
+                        <th style="width: 110px;"><label for="loginEmail">이메일</label></th>
                         <td style="width: 250px;"><input type="text" class="form-control" id="loginEmail" name="email"
                                                          placeholder="아이디를 입력하세요."></td>
                     </tr>
                     <tr>
-                        <td><label for="loginPassword">비밀번호</label></td>
+                        <th><label for="loginPassword">비밀번호</label></th>
                         <td><input type="password" class="form-control" id="loginPassword" name="password"
                                    placeholder="비밀번호를 입력하세요."></td>
                     </tr>
@@ -126,6 +126,7 @@
 
     });
 
+    // 회원가입 검증
 
     // 회원가입 버튼 비활성화
     $('#signUpBtn').attr("disabled", true);
@@ -137,36 +138,22 @@
     var pwd2 = $("#checkPW").val();
 
     // 비밀번호 글자 수 검증
-    $('#signUpPassword').focusout(function(){
-        if(pwd1.length<6 || pwd1.length>16){
-            alert("비밀번호는 6자리 이상, 16자리 미만으로 사용 가능합니다.");
-        }
-    });
-
-    // 비밀번호 확인
-    // if(pwd1.length==pwd2.length){
-    //     if($('').val().equals($('').val()){
-    //         availableSignUp = true;
-    //         // 조건 충족시 회원가입 버튼 활성화
-    //         if(availableSignUp){
-    //             $('#signUpBtn').removeAttr("disabled");
-    //         }
-    //     } else {
-    //         alert("비밀번호가 일치하지 않습니다. 비밀번호를 확인하고 다시 입력해주세요.");
-    //         $("#alert-success").css('display', 'none');
-    //         $("#alert-danger").css('display', 'inline-block');
+    // $('#signUpPassword').focusout(function(){
+    //     if(pwd1.length<6 || pwd1.length>16){
+    //         alert("비밀번호는 6자리 이상, 16자리 미만으로 사용 가능합니다.");
     //     }
-    // } else {
-    //     alert("비밀번호가 일치하지 않습니다. 비밀번호를 확인하고 다시 입력해주세요.");
-    //     $("#alert-success").css('display', 'none');
-    //     $("#alert-danger").css('display', 'inline-block');
-    // }
+    // });
 
     // 비밀번호 재확인 일치 여부
     $('#checkPW').focusout(function () {
 
         var pwd1 = $("#signUpPassword").val();
         var pwd2 = $("#checkPW").val();
+
+        console.log(pwd1+" : "+pwd2);
+        console.log(pwd1.length+" : "+pwd2.length);
+        console.log("boolean : "+(pwd1.length==pwd2.length));
+
         if (pwd1 != "" || pwd2 != "") {
             if (pwd1 == pwd2) {
                 $("#alert-success").css('display', 'inline-block');
@@ -184,6 +171,7 @@
             }
         }
     });
+
 
 </script>
