+++ src/main/resources/templates/layout/main.html
@@ -436,28 +436,24 @@
             
             
           </div>
-          <div class="col-lg-4 mb-5 mb-lg-0">
-            <div class="row mb-5">
-              <div class="col-md-12">
-                <h3 class="footer-heading mb-4">Navigations</h3>
-              </div>
-              <div class="col-md-6 col-lg-6">
-                <ul class="list-unstyled">
-                  <li><a href="#">Home</a></li>
-                  <li><a href="#">Buy</a></li>
-                  <li><a href="#">Rent</a></li>
-                  <li><a href="#">Properties</a></li>
-                </ul>
-              </div>
-              <div class="col-md-6 col-lg-6">
-                <ul class="list-unstyled">
-                  <li><a href="#">About Us</a></li>
-                  <li><a href="#">Privacy Policy</a></li>
-                  <li><a href="#">Contact Us</a></li>
-                  <li><a href="#">Terms</a></li>
-                </ul>
-              </div>
-            </div>
+<!--          <div class="col-lg-4 mb-5 mb-lg-0">-->
+<!--            <div class="row mb-5">-->
+<!--              <div class="col-md-12">-->
+<!--                <h3 class="footer-heading mb-4">Navigations</h3>-->
+<!--              </div>-->
+<!--              <div class="col-md-6 col-lg-6">-->
+<!--                <ul class="list-unstyled">-->
+<!--                  <li><a href="#">Home</a></li>-->
+<!--                  <li><a href="#">Buy</a></li>-->
+<!--                </ul>-->
+<!--              </div>-->
+<!--              <div class="col-md-6 col-lg-6">-->
+<!--                <ul class="list-unstyled">-->
+<!--                  <li><a href="#">About Us</a></li>-->
+<!--                  <li><a href="#">Contact Us</a></li>-->
+<!--                </ul>-->
+<!--              </div>-->
+<!--            </div>-->
 
 
           </div>
